ml_forecast/requirements.txt
# ML и данные
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Базы данных и API
psycopg2-binary==2.9.6
requests==2.31.0
sqlalchemy==2.0.19

# Утилиты
# ML и данные
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Базы данных и API
psycopg2-binary==2.9.6
requests==2.31.0
sqlalchemy==2.0.19

# Утилиты
python-dotenv==1.0.0
joblib==1.3.1

# Логирование
structlog==23.1.0

# Тестирование (опционально)
pytest==7.4.0

# API
fastapi==0.103.1
uvicorn==0.23.2